use atomicu64

+++ src/collector.rs
@@ -1,4 +1,4 @@
-use std::{thread, sync::{Mutex, atomic::{AtomicUsize, Ordering}}, collections::HashMap, cell::RefCell};
+use std::{thread, sync::{Mutex, atomic::{AtomicU64, Ordering}}, collections::HashMap, cell::RefCell};
 use slog::Drain;
 use super::{event::Event, span::Span};
 
@@ -42,8 +42,8 @@ pub struct Collector {
   spans: Mutex<HashMap<tracing::Id, Span>>,
   traces: Mutex<HashMap<u64, Vec<tracing::Id>>>,
   current: CurrentSpanPerThread,
-  span_id: AtomicUsize,
-  trace_id: AtomicUsize,
+  span_id: AtomicU64,
+  trace_id: AtomicU64,
   logger: slog::Logger,
   dd_client: datadog_apm::Client
 }
@@ -59,8 +59,8 @@ impl Collector {
       level,
       spans: Mutex::new(HashMap::new()),
       traces: Mutex::new(HashMap::new()),
-      span_id: AtomicUsize::new(1),
-      trace_id: AtomicUsize::new(1),
+      span_id: AtomicU64::new(1),
+      trace_id: AtomicU64::new(1),
       current: CurrentSpanPerThread::new(),
       logger: logger.new(o!()),
       dd_client: datadog_apm::Client::new(config)
@@ -78,12 +78,12 @@ impl tracing::Subscriber for Collector {
     let trace_id = parent.as_ref().map(|parent_id| 
       spans.get(parent_id).map(|parent_span| parent_span.trace_id)
     ).flatten().unwrap_or_else(|| {
-      self.trace_id.fetch_add(1, Ordering::SeqCst) as u64
+      self.trace_id.fetch_add(1, Ordering::SeqCst)
     });
     let mut traces = self.traces.lock().unwrap();
     let trace_spans = traces.entry(trace_id)
       .or_insert_with(|| vec![]);
-    let span_id = self.span_id.fetch_add(1, Ordering::SeqCst) as u64;
+    let span_id = self.span_id.fetch_add(1, Ordering::SeqCst);
     let span_id = tracing::Id::from_u64(span_id);
     let span = Span::new(parent, trace_id, span);
     spans.insert(span_id.clone(), span);
